back/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
numpy==2.2.5
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.2.3
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
sqlparse==0.5.3
threadpoolctl==3.6.0
uritemplate==4.1.1
urllib3==2.4.0
